--- app/src/main/java/com/example/weather/network/City.kt
@@ -1,14 +1,0 @@
-package com.example.weather.network
-
-data class City(
-    val clouds: Clouds,
-    val coord: Coord,
-    val dt: Int,
-    val id: Int,
-    val main: Main,
-    val name: String,
-    val sys: Sys,
-    val visibility: Int,
-    val weather: List<Weather>,
-    val wind: Wind
-)
