--- app/src/main/java/com/example/weather/network/Clouds.kt
@@ -1,5 +1,0 @@
-package com.example.weather.network
-
-data class Clouds(
-    val all: Int
-)
