--- app/src/main/java/com/example/weather/network/Coord.kt
@@ -1,6 +1,0 @@
-package com.example.weather.network
-
-data class Coord(
-    val lat: Double,
-    val lon: Double
-)
