--- app/src/main/java/com/example/weather/network/Sys.kt
@@ -1,8 +1,0 @@
-package com.example.weather.network
-
-data class Sys(
-    val country: String,
-    val sunrise: Int,
-    val sunset: Int,
-    val timezone: Int
-)
