--- app/src/main/java/com/example/weather/network/Weather.kt
@@ -1,8 +1,0 @@
-package com.example.weather.network
-
-data class Weather(
-    val description: String,
-    val icon: String,
-    val id: Int,
-    val main: String
-)
