+++ app/src/main/java/com/example/weather/network/WeatherApiService.kt
@@ -5,6 +5,7 @@ import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.GET
+import retrofit2.http.Query
 
 private const val BASE_URL = "https://api.openweathermap.org/"
 
@@ -28,8 +29,12 @@ private val retrofit = Retrofit.Builder()
  * A public interface that exposes the [getCities] method
  */
 interface WeatherApiService {
-    @GET("data/2.5/group?id=4772354,5368361,5128581,4887398,4699066,5809844,5856195,5419384,5391959,4684888,5746545,4990729,4335045&appid=bd4472d97ca0b479dc32513cf50a1bf3&units=imperial.json")
-    suspend fun getCities() : WeatherApiResponse
+    @GET("data/2.5/group?id={cityIds}&appid={appId}&units={units}.json")
+    suspend fun getCities(
+            @Query("cityIds") cityId: String,
+            @Query("appId") appId: String,
+            @Query("units") units: String
+    ) : WeatherApiResponse
 }
 
 /**
